[ENV] 필요없는 스크립트 삭제
