--- ProjectNT/Assets/03.Code/Scripts/XRDevice/StickTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d8d5d4a53429da4eae9bb1702a6e1bb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
