+++ ProjectNT/Assets/03.Code/Scripts/XRDevice/XRDeviceBase.cs
@@ -8,8 +8,8 @@ public class XRDeviceBase : MonoBehaviour
 {
     [SerializeField] private InputActionReference menu_BTN;
     //스틱은 컨트롤러의 자식 오브젝트
-    private StickTest leftStick;
-    private StickTest rightStick;
+    // private StickTest leftStick;
+    // private StickTest rightStick;
 
     public XRBaseController leftHand;
     public XRBaseController rightHand;
@@ -26,8 +26,8 @@ public class XRDeviceBase : MonoBehaviour
     {
         //스틱이 생성되는 걸 기다리기 위한 한프레임 휴식
         yield return null;
-        leftStick = leftHand.model.GetComponent<StickTest>();
-        rightStick = rightHand.model.GetComponent<StickTest>();
+        // leftStick = leftHand.model.GetComponent<StickTest>();
+        // rightStick = rightHand.model.GetComponent<StickTest>();
     }
 
     private void Menu(InputAction.CallbackContext context)
